+++ udata_front/theme/gouvfr/assets/js/components/AdminSidebar/AdminSidebarOrganizationMenu/AdminSidebarOrganizationMenu.vue
@@ -37,6 +37,11 @@
           :label="t('Members')"
           :to="{name: 'organization-members', params: {oid: organization.id}}"
         />
+        <AdminSidebarLink
+          icon="fr-icon-user-line"
+          :label="t('Profile')"
+          :to="{name: 'organization-profile', params: {oid: organization.id}}"
+        />
       </ul>
     </div>
   </li>