+++ udata_front/theme/gouvfr/assets/js/pages/OrganizationPublishingForm/Step1CreateOrJoinOrganization.stories.ts
@@ -3,7 +3,6 @@ import type { Meta, StoryObj } from '@storybook/vue3';
 import { within, waitFor, userEvent } from '@storybook/testing-library';
 import { expect } from '@storybook/test';
 import Step1CreateOrJoinOrganization from './Step1CreateOrJoinOrganization.vue';
-import * as Stepper from '../../components/Form/Stepper/Stepper.stories';
 
 const meta = {
   title: 'Pages/OrganizationPublishingForm/Step1',
@@ -16,9 +15,7 @@ const meta = {
 
 export default meta
 
-const args = {
-  steps: Stepper.StepperOrgaOnFirstStep.argsOrga.steps,
-};
+const args = {};
 
 export const Step1: StoryObj<typeof meta> = {
   render: (args) => ({
@@ -27,7 +24,9 @@ export const Step1: StoryObj<typeof meta> = {
       return { args };
     },
     template: ` <div class="bg-grey-50 fr-p-4w">
-                  <Step1CreateOrJoinOrganization v-bind="args" />
+                  <div class="fr-container">
+                    <Step1CreateOrJoinOrganization v-bind="args" />
+                  </div>
                 </div>`,
   }),
   args,
@@ -50,7 +49,9 @@ export const Step1WithInteraction: StoryObj<typeof meta> = {
       return { args };
     },
     template: ` <div class="bg-grey-50 fr-p-4w">
-                  <Step1CreateOrJoinOrganization v-bind="args" />
+                  <div class="fr-container">
+                    <Step1CreateOrJoinOrganization v-bind="args" />
+                  </div>
                 </div>`,
   }),
   args,